+++ src/pages/_app.tsx
@@ -1,5 +1,5 @@
-import { axiosRequestSuccessLambda } from '@/lambdas/axios-request-error.lambda'
-import { axiosRequestErrorLambda } from '@/lambdas/axios-request-success.lambda'
+import { axiosRequestSuccessLambda } from '@/lambdas/axios-request-success.lambda'
+import { axiosRequestErrorLambda } from '@/lambdas/axios-request-error.lambda'
 import axios from 'axios'
 import type { AppProps } from 'next/app'
 import { RecoilRoot } from 'recoil'
